fix Coverity findings

+++ src/daemon.c
@@ -521,7 +521,7 @@ int processifinfo(DSTATE *s, datacache **dc)
 		/* skip update if previous update is less than a day in the future */
 		/* otherwise exit with error message since the clock is probably messed */
 		if ((*dc)->updated > (ifinfo.timestamp + 86400)) {
-			snprintf(errorstring, 1024, "Interface \"%s\" has previous update date too much in the future, exiting. (%u / %u)", (*dc)->interface, (unsigned int)(*dc)->updated, (unsigned int)ifinfo.timestamp);
+			snprintf(errorstring, 1024, "Interface \"%s\" has previous update date too much in the future, exiting. (%" PRIu64 " / %" PRIu64 ")", (*dc)->interface, (uint64_t)(*dc)->updated, (uint64_t)ifinfo.timestamp);
 			printe(PT_Error);
 			errorexitdaemon(s, 1);
 		}
